refactor(game): split level logic into many files
